feat: SAGA transformer Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
